Rename mock_grains module to mock_grain
